updated notebook version to 729. added font selection cells.
text wrapping is hopeless. every fix I tried ended up adversely affecting the step by step building animation mode. apparently once a font reduction is triggered by a word not fitting its line within the cell, it keeps reducing to the minimum font even when not necessary. this is most likely because of the unpredictability of measuring the pixel length of a text against the chord length it sits on within a cell.

it is actually more predictable for texts bent along arcs because calculating the arc length is more consistent for every angle width of a cell. text bent along arcs is more computationally heavy though and needs rasterization so that the wheel stays nimble and frictionless.